@@ -20,7 +20,7 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >= 3.8
+python_requires = >= 3.9
 install_requires =
     numpy
     networkx
@@ -41,9 +41,6 @@ test =
 [options.entry_points]
 console_scripts = 
     openfe = openfecli.cli:main
-
-[bdist_wheel]
-universal = 1
 
 [versioneer]
 VCS = git
