@@ -1,8 +1,11 @@
 # This code is part of OpenFE and is licensed under the MIT license.
 # For details, see https://github.com/OpenFreeEnergy/openfe
 
+from . import _rbfe_utils
+
 from .equil_rbfe_methods import (
     RelativeLigandTransform,
     RelativeLigandTransformSettings,
-    RelativeLigandTransformResults,
+    RelativeLigandTransformResult,
+    RelativeLigandTransformUnit,
 )
