@@ -7,8 +7,8 @@ class TestAtomMapper:
     def test_abstract_error(self, simple_mapping):
         # suggest_mappings should fail with NotImplementedError if the user
         # tries to directly user the abstract class
-        molA = simple_mapping.molA
-        molB = simple_mapping.molB
+        molA = simple_mapping.componentA
+        molB = simple_mapping.componentB
         mapper = LigandAtomMapper()
         match_re = "'LigandAtomMapper'.*abstract.*_mappings_generator"
         with pytest.raises(NotImplementedError, match=match_re):
@@ -17,16 +17,16 @@ class TestAtomMapper:
     def test_concrete_mapper(self, simple_mapping, other_mapping):
         # a correctly implemented concrete atom mapping should return the
         # mappings generated by the _mappings_generator
-        molA = simple_mapping.molA
-        molB = simple_mapping.molB
+        molA = simple_mapping.componentA
+        molB = simple_mapping.componentB
 
         class ConcreteLigandAtomMapper(LigandAtomMapper):
             def __init__(self, mappings):
                 self.mappings = mappings
 
-            def _mappings_generator(self, molA, molB):
+            def _mappings_generator(self, componentA, componentB):
                 for mapping in self.mappings:
-                    yield mapping.molA_to_molB
+                    yield mapping.componentA_to_componentB
 
         mapper = ConcreteLigandAtomMapper([simple_mapping, other_mapping])
         results = list(mapper.suggest_mappings(molA, molB))
