@@ -88,15 +88,15 @@ class AtomMappingEdge(Edge):
         # figure out which node is to the left and which to the right
         xs = [node.xy[0] for node in self.node_artists]
         if xs[0] <= xs[1]:
-            left = mapping.molA
-            right = mapping.molB
-            left_to_right = mapping.molA_to_molB
+            left = mapping.componentA
+            right = mapping.componentB
+            left_to_right = mapping.componentA_to_componentB
+            right_to_left = mapping.componentB_to_componentA
         else:
-            left = mapping.molB
-            right = mapping.molA
-            left_to_right = {v: k for k, v in mapping.molA_to_molB.items()}
-
-        right_to_left = {v: k for k, v in left_to_right.items()}
+            left = mapping.componentB
+            right = mapping.componentA
+            left_to_right = mapping.componentB_to_componentA
+            right_to_left = mapping.componentA_to_componentB
 
         left_extent, right_extent = self._get_image_extents()
 
