@@ -69,7 +69,7 @@ def generate_mapping(mapper, molA, molB):
 def atommapping_print_dict_main(mapper, molA, molB):
     """Main function for generating and printing out the mapping"""
     mapping = generate_mapping(mapper, molA, molB)
-    print(mapping.molA_to_molB)
+    print(mapping.componentA_to_componentB)
 
 
 def atommapping_visualize_main(mapper, molA, molB, file, ext):
@@ -88,9 +88,9 @@ def atommapping_visualize_main(mapper, molA, molB, file, ext):
             "supported: " + ", ".join([f"'{ext}'" for ext in ext_to_artist])
         )
 
-    contents = visualization.draw_mapping(mapping.molA_to_molB,
-                                          mapping.molA.to_rdkit(),
-                                          mapping.molB.to_rdkit(), d2d=artist)
+    contents = visualization.draw_mapping(mapping.componentA_to_componentB,
+                                          mapping.componentA.to_rdkit(),
+                                          mapping.componentB.to_rdkit(), d2d=artist)
 
     file.write(contents)
 
